requirements.txt
opencv_python==4.2.0.32
Shapely==1.7.0
xmljson==0.2.0
thop==0.0.31.post2005241907
matplotlib==3.1.3
imgaug==0.4.0
scipy==1.4.1
p_tqdm==1.3.3
lxml==4.5.0
tqdm==4.43.0
ujson==1.35
PyYAML==5.3.1
scikit_learn==0.23.2
tensorboard==2.3.0
gdown==3.12.2
